reorganised api functions alphabetically, created file for general helpful functions
